Fixed duplicate characters
According to the Validator, the mod does no longer have any duplicate
characters in it. What I did was to add a "0" at the end of all numbers
that the Validator saw as duplicate. Hopefully this can be the first
step towards fixing the crashes and doing some general house-cleaning
around the history present in the mod.