Restoration requirements #16
I have modified the conditions of restoring the Legions. There is now a
pseudo-system in place. You have to control the home-duchy of the
Legion, but you have to be the dejure King of that duchy to do so.

I did this in correlation with the terms of the Reconquests, that you
have to be a King to do.

Not all Legions follow this, but the majority of them do.