Error Corrections, Wahhabism
1. I fixed my mistake of removing "always = no" from tribal titles.
2. I removed the features I added to the Mimamsa Hindu Heresy, per
Loup's request, because Autocephalacy will be fully editable in the next
Patch.
3.. Wahhabism is no longer a clone of Sunni Islam. Wahhabism is the sect
of ISL, Al-Qaeda, Saudi Arabia, and Islamic extremist terrorists the
world over. Wahhabis have brought Americ and the West to their knees
time and time again, and they have been like this since the 700s and
800s AD. Wahhabis are some of the best and most unconventional fighters
on the earth, and I made sure to simulate that!