Attila the Hun's Family Deserves Their Own Coat of Arms!
Let it be said that I have absolutely no idea how to code special coats
of arms for unique dynasties. I tried my best and I'd like you all to
fix it.

In this struggle, I tried to encode an unique Coat of Arms for  House
The Hun. Attila the Hun is one of the few antiquity "barbarians" that we
both remember and fear.

I think his legacy deserves a fitting Coat of Arms for those players who
wish to revive his legendary empire.

For inspiration, I used this picture:
https://upload.wikimedia.org/wikipedia/commons/0/0c/Nuremberg_chronicles_-_Atilla,_King_of_the_Huns_%28CXXXVII%29.jpg
.

Notice the sword and the scourge-whip. I included both in the COA.