Flag Improvements! (P1)
I tried to dent the overwhelming problem of duplicate, colorless, and
unoriginal placeholder flags in the GFX folder.

Here is an explanation of what I did to each, in order of the change
log:

1. C_CHAGAI - The red blob on the black center is simply an outline of
the borders of Chagai in RL.

2. C_KAZAKH - Intended to resemble the modern flag of Kazakhstan, which
in turn is also based off of an ancient Tengri motif. I tried to make it
look different than an existing province I recalled with a blue and
yellow color scheme.

3.  C_ZARANJ - A quick Google search of Zaranj yielded no results except
a map or two. For some reason, the German Flag showed up. I edited that
a bit to get this result. I tried not to make it look too much like
Germany's flag.

4. C_WIGHT - That's the actual flag of Wight.

5. D_HASUAWARPARTY - A quick search yielded the Hasua race and their
tradtional dress. The pattern is taken from that.

6. D_HORMUZ_REPUBLIC - I assumed that this was the title name for the
ancient Qarmatian Republic. Working on that, I Googled "Qarmatian
symbols" and got that epic calligraphy there. I cut off a portion of it
and made it into a nifty flag.

7. D_MAYOR_FRANKS - I Googled "Franks symbols". That was one of the
first that came up.

8. D_PLATONIC_ACADEMY - That's one of the first results when you Google
"Neoplatonism". It shows Plato himself. I almost went with the famous
image of Plato's stone head, but decided against it.

9. K_SOGDINIA - Sogdinia was a real place and a real kingdom. It was too
ancient to have any known flag, so I chose a stylized, ancient coin
minted by that kingdom to represent it. Cool, huh?

10. K_CRETE - That is a version of the actual flag of Crete. As the axe
motif was already duplicated elsewhere, I simply used this instead.

11. D_ZUN_WARRIORS - Fixed the error where the Knights of Calavatrava
had their logo in the Zunist Holy Order. I used the Vanilla banner
instead.