Flag Improvements! (P2)
Here we go again!

1. C_KIMAK - Googling "Kimak" yielded this bad-A symbol.
https://pbs.twimg.com/profile_images/1852434872/logo_400x400.png - why
not use it? It doesn't have any real cultural meaning.

2. Googling "Kangy" yielded nothing but kangaroos. I took a pattern from
one of the furry backgrounds and shrunk it to remove the kangaroos.

3. C_KYZYLKUM - Kyzylkum is a real place and a real desert. What you are
looking at is a real palace in that real desert. Specifically, this is
part of the complex of the walls of the City of Kiva. -
http://www.centralasiarally.com/2016/03/secrets-kyzylkum-desert/ - I'm
not sure if this is actually in Khiva. The alternate version is courtesy
of a filter called "Deep Dream", provided by www.ezimba.com . If the
palace is not in that area, we'll use it instead.

4.  C_OTRAR - In RL, Otrar is a ghost town that once hosted the Silk
Road. What you are looking at is the logo of Otrar Travel. Otrar Travel
is a Kazakh corporation that helps Kazakhs travel the world.

5. D_IBADI - The Ibadi Caliphate's flag no longer copies Sunni Islam.
You are looking at a RL inscription on the Ibadi Mosque of Oman's
capital.

6. D_ROMAN_PREFECT - FINALLY, that placeholder using an Afghan symbol is
GONE! The symbolism is simple and obvious. I chose two different
versions of the flag. One  is based on a Roman banner and the other is
based on the flag of the Vatican.