Flag Improvments (P4)
Before I say anything else, I just want to thank you for letting me do
this. This means so much to me. I tried to help the Game of Thrones Mod
with their blank flags, and they actually insulted me and laughed at the
work I spent hours on.
You, on the other hand, have embraced and incorporated my work into your
Mod. After being so degraded and insulted by the GOT overseers, it's
good to be appreciated.

On to the changes!

1. C_ARAL - I simply added currents of blue to represent the Aral Sea.

2. C_ILI - This simple flag represents the Ili River, with a tasteful
blue stripe.

3. C_SARAVAN - This beautiful flag came up when I searched "Saravana".
Saravana County is the district that has the modern city of Saravan as
its capital.

4. C_ZHETYSU - Did you know that this province's modern day equivalent
hosts a soccer team? What you are looking at is a modified version of
that team's logo. I removed the lines and stripes from the central
soccer ball (so that it looks like a circle with ornate grey spots) and
cleaned out the modern language wording. It looks great. I wish I knew
the size of the shields, though, because then I could match it to that
size.

5. D_ZHETYSU - See #4, but this one is PURPLE!

6. K_AL_HIRAH - This is an edit of an Arabic book cover documenting
Al-Hirah's history.