Windows Help
These are the support files for Windows Help. Due to copyright issues,
all installers are removed and replaced with a formal link to Microsoft,
per Loup's request. The actual Help File may be delayed until the Mod is
out of Beta in order to ensure a full, comprehensive feature list.