Kek's Meme Magic and Flag Improvements (P1)
Pardon the mess. I am teaching myself how to write Events, and I thought
this would be a great first chance. As I said I would, I created a
randomly occuring Event that happens in when an Egyptian Pagan chooses
to Suvey the Realm.

I will try and play test this Event, but I can already be certain that
it will fail. I may need help.

------

I also changed some flags. The explanations are as follows:

1. C_RQA - This is the logo for the Research Quality Association (RQA),
with all words and branding removed. Nothing else came up for RQA, so I
just used the stylized curves for it. It looks pretty.

2. D_JOUF - Jouf is a Saudi province and is home to Sakaka City. In
Sakaka is a giagantic novelty teapot. You are looking at a render of
that teapot. Jouf has no actual flag in RL.

3. D_MAKKAH - This is the flag of the proposed Province of Islamabad in
Pakistan. It looks like a perfect encapsulation of Islam and Mecca in
general, without looking "too Muslim" if non-Muslims hold it.

4. D_TRANSOXIANA - This graphic is courtesy of www.transoxiana.org !
Transoxiana was home to a massive temple complex for pre-Islamic pagans.