Translate the loading tips in French #87
Some of them are not translated because:
- Some of the peoples are not translated
- I have not understand some them

I have also noted that the localisation of the varangian trait is still
the same as the Vanilla. In fact the trait is used both for Varangian
Guard, Scholae Palatinae and Excubitors. I think we need to create new
traits for the new guards. Another option is to have neutral text for
this trait.