Flag Improvements! (P3)
1. C_BORKOU - Borkou is a province of the nation of Chad. What you are
looking at is the flag of Chad with the borders of modern Borkou on it.

2. C_BSLAMA - "Bslama" means "goodbye" according to Google. You are
seeing a symbol I found when I googled "bslama".

3. C_BENI_ISGUEN - What you are looking at is a stylized, condensed
rendition of the suburbs of modern Beni-Isguen. When shrunk to a 128x128
size, it looks like a pattern instead. I simply crossed it with a
diaganol line to complete the image.

4. C_BOUNOURA - Every other picture that appeared when I Googled this
province yielded the Great Mosque of Bou Noura. Much to my surprise, it
was the same building that was on the flag of Mzab in EU4.

5. C_CHUY - Every other search result for "chuy" was "CHUY FOR MAYOR
CHUY FOR MAYOR". Chuy ran for the mayor of Chicago (no idea why anyone
would want to go near that cesspit), so I decided to borrow aspects of
Chicago's flag (the only thing good about that tumor of a city) for the
province.

6. C_FADA - Almost every single search result for "fada" showed fairies.
I chose a clipart piece and added it to the  flag.

7. C_JASK - This is a modified version of Jask Property, Limited, but
without the corporate branding or copyrighted items.

8. C_KALAT - Actual flag of the Khanate of Khalat, founded in 1666 AD.
No earlier flag exists.

9. C_MAKRAN - Actual flag of the Princely State of Makran.

10. C_OSHRUSANA - Oshrusana is a place apparently connected to a man
named
Khaydhar ibn Kawus al-Afshin. This is a tapestry connected to said man.