Flag Improvements! (P4)
1. D_TRANSOXIANA moved to K_TRANSOXIANA

2. D_TRANSOXIANA - Edited version of "Vampire TranSOXiana" by "Goblyn's
Comics / Keith Churchill",
http://goblynscomics.com/20120717_vampire_sox.htm

3. K_VATSAGULMA - Vatsagulma is apparently an area near Kashmir. This
statue comes from
https://imgv2-1-f.scribdassets.com/img/document/222540346/original/1d25daafac/1464754927
and its ancient Buddha statues.

4. K_VISIGOTHS - The creepy bird you see there came up constantly when I
Googled Visigothic symbols and flags.

5. K_PADISHKHWARGAR - This kingdom was apparently a vassal state of the
Sassanids. What you see here is an alternate symbol for the Sassanid
Empire.

6. K_THESSALONIKA - Actual RL flag of the City of Thessalonika! Sadly,
it's a modern flag, but it'll have to do.

7. K_TRIPOLITANIA - Actual flag of Tripolitania in Libya!

8. K_THURINGIA - As boring as this is, this is unfortunately the true
flag of Thuringia, except with its coat of arms.

9. K_MASMUDA - This is the flag of the Almohad Kingdom in Morrocco.
Supposedly, the person that rose up and united the disparate Masmuda
Tribes was from here, and he formed the Almohad Kingdom.

10. K_LOMBARDIA - To differentiate the Kingdom of LombardIA from the
Kingdom of LombardY, I have opted to use a real flag of Lombardy dating
to the 1990s AD, superimposed with the ancient snake that the Lombards
are known for. I'm not sure why there are two Lombard Kingdoms in this
Mod, but this flag is intended for the SECONDARY kingdom. The primary
will use the snake eating that guy.

11. E_TAMIL - Actual flag of the Empire of Pallava, which ruled over all
of Tamil during this Mod's actual time period! "The Pallava royal
insignia was the Simha (lion) and Nandi (bull) which was changeable.
Narashimavarman used the lion as his emblem and Nandivarman preferred
the Nandi. Under the reign of Parameswaravarman was Khatvanga (Skull
mace) added to Pallava emblem."

12. K_BELGICA - This is a drawing of Leo Belgicus, which has long been
used in Belgium's heraldy to represent the country and its regions. The
real name for Belgica was "Gallia Belgica". This is far better than the
modern Belgium flag without any changes to make it fit.