Flag Improvements! (P5)
1. E_KUSHAN_EMPIRE - This is a statue of the head of a Kushan Prince. As
the Kushan Empire had no flag, this is the best I could get. -
https://en.wikipedia.org/wiki/Kushan_Empire#Early_Kushans

2. K_AMR_CONFEDERACY - This flag is taken from a painting of the Battle
of the Trench, where the Amr Confederacy faced off against the forces of
darkness. The person holding the severed leg is Amr ibn ‘Abd Wudd
himself, the leader and namesake of the Amr Confederacy.

3. E_FRANKS - This is the Oriflamme. It's the flag of Charlemagne and
his army. It is the oldest known flag of what we now call France.