The Struggle is Real
OKAY, so I got the thing to fire! Now I need to get the option that
ISN'T the cynical one to show up alongside that. -AND I need to get the
second event (which now fires three times in a row) to not only show its
proper option, but remain triggered only once.