Flag Improvements (P6) and Frog Improvements!
1. C_BALKASH - This is a representation of a large telecommunications
tower on the lakeshore of the primary city/seat of Balkash. The blue is
for Tengri and water.

2. C_CHACH - Chach is an ancient name for the current capital city of
Uzbekistan, the City of Tashkent. I simply took portions of the Coat of
Arms/Seal of Tashkent and placed them here.

3. C_DASHOWUZ - Rendering of purses for sale in Dashowuz-region markets.
This, as a bonus, shows off famous Turkmen weaving.

4. C_DAWKAH - Dawkah is famous for its frankincense. What you are
looking at is  a health advertisement featuring a thumbnail of
frankincense surrounded by attractive backgrounds. /
http://aromatherapyhealthyessentialoils.com/wp-content/uploads/2014/03/Frankincense-e1437021969828.png

5. C_EL_ATEUF - El Alteuf is a historic region of the Mzab region. You
know that that meanse: MORE RENDITIONS OF MOSQUES SHAPED LIKE THE HAT OF
FINN FROM "ADVENTURE TIME"!

6. C_KATTEGAT_ISLANDS - The Kattegat region is home to the property of
famous viking Ragnor Lobrok. I simply edited a shirt with his name and
Viking weaponry on it to remove branding and English lettering. /
https://img0.etsystatic.com/045/0/9659731/il_570xN.737888000_j501.jpg

----

All dank frog-related Localization has been consolidated in a central
frog zone. It feels good, man.