Flag Improvements! (P7)
1. C_KET - The cover of the Altaic Origins book shown at
http://s155239215.onlinehome.us/turkic/20Roots/201Altaians/Potapov-TeleEn.htm
, without the text, flipped upside down

2. C_TIS - A gloved hand. I had nothing else.

3. C_ZAHEDAN - Zahedan is a region of Iran. This is an image I found
while searching for that name.

4. K_SEMIEN - Replaced the generic Star of David with the EU4 Semien
flag, using the colors of that region of Ethiopia

5. C_URZHAR - This is a topographic map of the RL Urzhar region.

6. C_TURGAY - This is the Coat of Arms of the Russian Empire's Turgay
Oblast!

7. C_TURA  - These are ancient Turkic sculptures. Though unconnected to
Tura itself, they nonetheless manifested in that region.

8. C_SYR_DARYA - A beautiful representation of the Syr Darya River.

9. C_SISTAN - Scaled down, stylized representations of coins from
Sakastan. Sakastan is another term used for Sistan.

10. C_T_UR_U - Let me be very clear on this one. I have truly suffered
to make this Mod great. A Google search of "turu" yielded pages of
disgustingly naked women. I grabbed the first symbol I saw and ran the
heck out of there.

11. D_JUDHAM - The Banu Judham was a tribe in opposition to Muhammed at
the dawn of Islam. This is a symbol associated with Jabal Amel, where
they ultimately settled.

12. D_TURIN - The image of a bull rearing back is the official symbol of
Turin, and has been so for ages. (The secondary image is the official
Flag of Turin. Notice the bulls.)