Flag Improvements! (P1)
1. C_IRAM - Artists' rendition of a major city in the Iram region. (IIRC
from yesterday, it's a Persian area.)

2. D_CYBELIAN_PAGAN - You are looking at a stylized clipart of a
pinecone. Why? Pine cones were sacred to the Cybelian Mysteries and
considered a symbol of regeneration, renewal, and fertility. In fact,
there exists a giant pinecone near the Vatican, flanked with two bronze
birds, that was created by a Cybelian who worshipped right under the
Pope's nose. (The alternate flag is a sacred castration tool, used to
painfully produce eunuchs for male clergy of that faith.)

3. D_DUBAI - Actual flag of Dubai. It's all I  could find.

4. D_GAETULIA - This is the flag used for Gaetulia by the EU4 "Extended
Timeline" Mod.

5. D_GURMA - This corporate logo has some sort of  connection to the
Gourma (or Gurma) culture.

6. D_HADRAMAWT - In RL Hadramawt, women famously wear black with "Wander
Over Yonder"-esque hats that make me think of Doctor Seuss. What you are
looking at is a picture of that hat.

7. D_HHAHA - The Haha (or Hhaha) peoples are a Berber tribe. You are
looking at a Berber symbol.

8. D_HISPALENSIS - One recurring theme that appeared when I Googled this
location was "ISIDORE OF SEVILLE", a Catholic saint. from the 500s AD He
successfully converted the Visigoths to Nicene Christianity from
Arianism. I chose a symbol representing him for use as a flag.

9. D_KAAB - This is a modified corporate logo from a company catering to
a region where the Kaab peoples are found.

10. D_KALB - A Google search of "Kalb" constantly yielded pictures of
cows and pictures of calfs. I theorize that "Kalb" means "calf" or "cow"
in some foreign language. The background was originally black with a
bright green "K". As the Latin alphabet was not used here, I changed the
"K" a bunch of radiating lines from a central point and placed a calf
over it.

11. D_KAMANDARAN - This is a specifically appealing Arabic-esque picture
I found when Googling that name.

12. D_KANURI_WARPARTY - Actual flag of the Kanuri peoples.

13. D_LAGUANTAN - A Berber/Maghreb coin from the time period covered in
this Mod has been placed on the flag.

14. K_NAPOLI - Actual and earliest known flag of Napoli. (The alternate
flag is the actual and modern flag of Naples/Napoli.)

In addition, I applied some filters to increase the Sharpness of the
flag of Salem in order to make it more discernible and less blurry
in-game.