Changes
1. The King of Oriens now has a proper FOA.

2. Gepids can now restore Gepida.

3. If Oriens becomes independent of Byzantium, they can elect to
establish their own Empire and shape their own destiny.

4. Steppe nomads of the right culture can now form their culture's
titular empire.

5. Fixed bizarre instance where a Central African character can appoint
a Griot, but cannot spawn one for that appointment to be made.

6. Teeing the code up for implementation of the unfinished Barborites,
which were a Gnostic religion present during this time period and on
into the mid-to-late 500s AD. Moved Barborites to main religious
localization sector pending implementation.

7. Fixed multiple spelling, grammar, punctuation, and other errors in
localization.