Borborites!
I am struggling with finding out why there is stray code in the
Decisions system (AI_WILL_DO, MODIFIER, etc.), especially because my
added Decisions actually work.

Also, I request that I not be credited for implementing the Borborites.
I would like to remain anonymous due to the grossly immoral and unclean
nature of "The Filthy Ones". Due to my own religious beliefs, I'd prefer
not to have my legacy on this Mod connected to it.