Fixed Decisions Issue and Traits
1. The issue where strange leftover code kept showing up after I added
the ability for Borborites to "employ" a shrine prostitute is fixed!

2. Realism and balance to several traits.

3. Inheriting left handedness is rare in RL, but ten percent of the
earth ends up with it as a circumstance of their birth. Thus, it has a
high spawn rate and a low chance to inherit it.

4. Being tall does not increase your health (I know this from
experience)., and if both your parents are tall, you are almost always
tall (again, this is from experience.

5. One's voice is genetically inherent and, alongside one's accent, is
passed along. If both of your parents have a melodious or powerful
voice, you are likely to get it yourself.

6. The code had several compliments and insults for different traits,
but they were not implemented. I did this for you.

7. Added missing health modifier for wrymouth.