Yazidi Relocation and Adoptionism Fix
Yazidis are now moved to the Zoroastrian area and, temporarily, will be
listed as a Zoroastrian heresy instead of an Islamic one.

Finally added omission of Kurdish and Yazidi communities in modern Syria
and Iraq. This was LOOOOOOOONG overdue.

Adoptionism now has its own head instead of the Appollinarian Chuirch
getting two heads that would crash the game if both existed.