Flag Improvements! - I
1. K_CAUCASIAN_ALBANIA -  ACTUAL FLAG OF CAUCASIAN ALBANIA (under the
Mihranids, who actually ruled during this time)

2. K_HUNAS -  Inscription of war scene of the Hunas, circa 500 AD

3.  K_MOREA  - EU4/EU4-ET flag of Morea/Achaea

4. C_ASSOUNGHA - Regional, African tribal art from Google Images when I
searched the term

5. C_AZIYRA - Portion of a marketing advertisement declaring the titular
area to be a "Culinary Nirvannah"

6. C_AYEZDUS - Screenshot from an Arabic brochure

7. C_BAQA - Logo for Baqa News, edited to remove trademarks and such

8.  C_GUERRA - "Guerra" is Spanish for "war". I chose a generic picture
of crossed swords.

9. C_HLESEY - Hlesey is an island in Denmark. For some reason, a lot of
pictures of chickens came up when I Googled that location. Since
chickens have appeared on heraldy anyway, I went with one of the
results.

10. C_SEM - This is apparently the logo of "SEM Zuaba Corp". All Google
results for "Sem" came back with a province in Cambodia.