Sweeping Improvements
1. Orthodox JEWS can now build JEWISH Infantry Barracks

2. Anyone can spawn war elephants from the special Building if they
reside on the Indian continent, where such a thing exists

3. Berbers can now build Berber Nomad stables and other Berber buildings
(why was this missing)

4. Assyrians and the Emperor of Assyria can now spawn camel calvaries,
as the Assyrian Empire was legendary in its mastery of the camel

5. Tzadikites, being a warrior religion focused on aggression, can now
raid/loot, and receive some bonuses to home units

6. Added bonuses for Assyrian Pagans

7. Added Babylonian Ashurism (already in code but not fully implemented)
as a heresy to the mainstream Assyrian religion. This was the religion
of the Assyrians under Babylonian rule and was heavilly syncretized with
the Babylonian pantheon, aside from Ashur being the chief deity. They
are far less aggressive than the true Assyrian Pagans.

8. JEWS can now spawn JEWISH Retinues

9. Assyrian Paganism is now listed as an "ancient religion", as it was
such in RL

10. Shuffled the nature of some faiths: the Egyptians were once one of
the largest religions in the Middle East because of their
prolestyzation, so they go to the Prolestyzation section. Assyrian
Pagans were equally zealous and aggresive in spreading their faith. Jews
do not prolestyze in RL, but do allow for Gentiles to convert, so they
were moved elsewhere. Melchisidechianists and Sethians, being persecuted
and zealous in themselves, belong in the Communitarian Trigger.

11. Mystics now only need eight Learning to join the Hermetic Society.
Assyrian and Babylonian Pagans can now join the Hermetic Society,
because their religion (especially the latter's) heavilly depended on
and worshipped the arts of astrology and alchemy. Babylonian Pagans even
called their clergymen "investigators".