Mystic Trait Alterations
1. Removed the Mystic Trait's designation as a "lifestyle". Many
characters of the Tzadikite and Melchisidechian religions (including the
player) receive the trait as a part of their office, making it
impossible for them to experience the Focus Events for other activities.
The alternative was raising the Lifestyle Traits Limit in dozens of
Events.

2. Added Event to prevent toddlers from being subjected to the Event
that gives characters of some religions the Mystic Trait