Add 24 new legions
These are the legions shown on the map in the forum (https://forum.paradoxplaza.com/forum/index.php?threads/list-of-createble-legions-and-their-attached-provinces.889354/#post-22075320) that weren’t yet in the mod. I understand the map is a work in progress, but should a legion’s home duchy move it’s a one line change to WtWSMS_organise_legions_decisions.txt and WtWSMS_legion_events.txt to update it.